refactoring in preparation for json_get_str
