+++ src/common_macros.rs
